Remove binarised anims, Separate Turret files.
