+++ addons/arcadian/base/config_turret_gatling.hpp
@@ -1,10 +1,9 @@
-class CargoTurret;
 class Turrets: Turrets {
     class MainTurret: MainTurret {
         body = "mainTurret";
         gun = "mainGun";
         viewGunnerInExternal = 1;
-        minElev = -30;
+        minElev = 30;
         maxElev = 45;
         initElev = 0;
         soundServo[] = {"", db-40, 1};
@@ -44,33 +43,5 @@ class Turrets: Turrets {
             initFov = 0.75;
         };
         class ViewGunner: ViewOptics {};
-    };
-
-    class CargoTurret_1: CargoTurret {
-        minElev = -10;
-        maxElev = 23;
-        maxTurn = 130;
-        minTurn = 25;
-        gunnerAction = "passenger_flatground_2";
-        gunnerName = "$str_a3_rear_gunner";
-        proxyIndex = 6;
-        LODTurnedIn = 1000;
-        LODOpticsIn = 1000;
-        LODTurnedOut = 1000;
-        LODOpticsOut = 1000;
-        isPersonTurret = 1;
-        memoryPointsGetInGunner = "pos cargo rear";
-        memoryPointsGetInGunnerDir = "pos cargo dir rear";
-        memoryPointGunnerOptics = "";
-        gunnerOutOpticsModel = "";
-        gunnerOpticsModel = "";
-        enabledByAnimationSource = "reardoor_2_source";
-        usepip = 0;
-        ejectDeadGunner = 1;
-        startEngine = 0;
-        outGunnerMayFire = 1;
-        inGunnerMayFire = 0;
-        commanding = -2;
-        gunnerCompartments = "Compartment1";
     };
 };
